Move datatables into their own file
