+++ core/views/__init__.py
@@ -5,3 +5,4 @@ from .record_group import *
 from .job import *
 from .record import *
 from .transformation_scenario import *
+from .datatables import *
